feat: added test, refactored delegate cmd
